% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{binning}
\alias{binning}
\title{Equidistant binning of spectra}
\usage{
binning(X, ppm, width = 0.001, npoints = NULL)
}
\arguments{
\item{X}{NMR matrix, rows represent spectra.}

\item{ppm}{Matching ppm vector.}

\item{width}{Width of bins}

\item{npoints}{Desired number of data points in binned ppm vector}
}
\value{
Returned is a matrix with spectra in rows and binned ppm variables in columns. Column names represent binned ppm variables.
}
\description{
Equidistant binning of spectra
}
\author{
Torben Kimhofer \email{tkimhofer@gmail.com}
}
