% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibration}
\alias{calibration}
\title{Spectral calibration to a chemical shift reference}
\usage{
calibration(NMR, ppm, type = "Urine")
}
\arguments{
\item{NMR}{Input NMR matrix with rows representing spectra and columns representing chemical shifts.}

\item{ppm}{ppm vector with the same length as \code{ncol(X)}.}

\item{type}{Either 'Urine' or 'Plasma' for urine or blood-derived spectra, respectively (see Details).}
}
\value{
Returned is the calibrated NMR data matrix.
}
\description{
Spectral calibration to a chemical shift reference.
}
\details{
Spectral calibration to a chemical shift reference. If \code{type='Urine'} calibration will be performed using signal resonating around 0 ppm (Trimethylsilylpropanoic acid resonance). If \code{type='Plasma'}, the doublet around 5.23, originating from the alpha anomer of glucose, will be used for calibration. \strong{Blood serum-derived spectra} can also be calibrated with input argument \code{type='Plasma'}.
}
\references{
Dona, A.C., \emph{et al.} (2014) Precision high-throughput proton NMR spectroscopy of human urine, serum, and plasma for large-scale metabolic phenotyping. \emph{Analytical Chemistry}. 86.19. 9887-94.
}
\author{
Torben Kimhofer
}
