% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dummy_Y.R
\name{create_dummy_Y}
\alias{create_dummy_Y}
\title{Create dummy matrix and check numeric variable for OPLS analysis}
\usage{
create_dummy_Y(Y)
}
\arguments{
\item{Y}{Dependent variable (numeric or factor for regression or discriminat analysis, resepctively)}
}
\value{
List of two: Dummy matrix and data frame mapping Y-levels to numeric representations.
}
\description{
Create dummy matrix and check numeric variable for OPLS analysis
}
