% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_sets_method.R
\name{cv_sets_method}
\alias{cv_sets_method}
\title{Generate k cross-validation sets}
\usage{
cv_sets_method(cv.k = 7, Y, type = c("R", "DA"),
  method = "k-fold_stratified")
}
\arguments{
\item{cv.k}{Number of cross validation sets (k-fold).}

\item{Y}{Dependent variable in vector or dummy matrix format (see Details).}

\item{type}{Indicating if a regression (\code{R}) or discriminant analysis (\code{DA}) will be performed.}

\item{method}{Type of cross validation: k-fold or Monte Carlo-Cross Validation (MCCV), sampling can be performed totally random or group stratified: 'k-fold', 'k-fold_stratified', 'MC', 'MC_stratified'.}
}
\description{
Generate k cross-validation sets
}
\details{
If input argument \code{Y} is a dummy marix then function \cite{\code{centre_scale}} has been applied beforehand.The number of cross validation sets is related to the number of samples in each group. If in doubt, set \code{k=nrow(Y)} or \code{k=length(Y)} if Y is matrix or vector, respectively.
}
\seealso{
center_scale
}
\author{
Torben Kimhofer \email{tkimhofer@gmail.com}
}
