% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmodx.R
\name{dmodx}
\alias{dmodx}
\title{Calculating distance to the model in X space}
\usage{
dmodx(model, plot = T)
}
\arguments{
\item{model}{OPLS model of type \code{OPLS_MetaboMate}.}

\item{plot}{Logical indicating if results should be visualised.}
}
\value{
The projection distance of each observation in the model (\code{DModX}).
}
\description{
Calculating distance to the model in X space
}
\references{
Bylesjö M., \emph{et al.} (2002) OPLS discriminant analysis: combining the strengthsof PLS-DA and SIMCA classification. \emph{Journal of Chemometrics}, 20, 341-51.

Wold S. (1976) Pattern recognition by means of disjoint princi-pal components models.  \emph{Pattern Recognition}, 8, 127-39.
}
