% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matspec.R
\name{matspec}
\alias{matspec}
\title{Plotting multiple NMR spectra}
\usage{
matspec(ppm, X, shift = c(0, 9.5), add = F, ...)
}
\arguments{
\item{ppm}{ppm vector.}

\item{X}{NMR matrix with spectra represented in rows.}

\item{shift}{Chemical shift region to be plotted (in ppm).}

\item{add}{Logical indicating if spectra should be added to a current plot generated with \code{spec()} or \code{matspec()}.}

\item{...}{Additional parameters to be passed on to the plot function.}
}
\description{
Plotting multiple NMR spectra
}
\details{
Low-level plotting function for NMR spectra.
}
\seealso{
\code{\link{spec}} \code{\link{plot}}
}
