% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minmax.R
\name{minmax}
\alias{minmax}
\title{Min-Max scaling}
\usage{
minmax(x)
}
\arguments{
\item{x}{Input vector.}
}
\value{
Scaled x vector.
}
\description{
Min-Max scaling
}
\details{
Scaling minimum and maximum values to zero and one
}
