% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIPALS_OPLS_component.R
\name{NIPALS_OPLS_component}
\alias{NIPALS_OPLS_component}
\title{Calculating a single OPLS component}
\usage{
NIPALS_OPLS_component(X, Y)
}
\arguments{
\item{X}{Input matrix with rows and columns representing observations and variables}

\item{Y}{Dependend variable, in form of dummy matrix (multi-levels allowed) or numeric column vector}
}
\value{
Returned is a list with the following entries:
\item{Filtered X}{Orthogonal filtered X matrix.}
\item{Scores X pred}{PLS component scores.}
\item{Loadings X pred}{PLS component loadings for X.}
\item{Weights pred}{PLS component variable weights for X.}
\item{Scores X orth}{Orthogonal component scores.}
\item{Loaadings X orth}{Orthogonal component loadings for X.}
\item{Weights X orth}{Orthogonal component X variable weights.}
\item{Loadings Y}{PLS component Y loadings.}
}
\description{
Calculating a single OPLS component. This function should only be used for two level-Y, coded as a numeric vector.
}
\references{
Trygg J. and Wold, S. (2002) Orthogonal projections to latent structures (O-PLS). \emph{Journal of Chemometrics}, 16.3, 119-128.
}
