% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIPALS_OPLS_component_multilevel.R
\name{NIPALS_OPLS_component_mulitlevel}
\alias{NIPALS_OPLS_component_mulitlevel}
\title{Calculate PLS component, orthogonalise X and remove orthogonal information}
\usage{
NIPALS_OPLS_component_mulitlevel(X, Y)
}
\arguments{
\item{X}{Input matrix with rows and columns representing observations and variables}

\item{Y}{Dependend variable, in form of dummy matrix (multi-levels allowed) or numeric column vector}
}
\value{
Returned is a list with the following entries:
\item{Filtered X}{Orthogonal filtered X matrix.}
\item{Scores X pred}{PLS component scores.}
\item{Loadings X pred}{PLS component loadings for X.}
\item{Weights pred}{PLS component variable weights for X.}
\item{Scores X orth}{Orthogonal component scores.}
\item{Loaadings X orth}{Orthogonal component loadings for X.}
\item{Weights X orth}{Orthogonal component X variable weights.}
\item{Loadings Y}{PLS component Y loadings.}
}
\description{
Calculate PLS component, orthogonalise X and remove orthogonal information. This function is not for general use, rather than part of the opls function.
}
\author{
Torben Kimhofer \email{tkimhofer@gmail.com}
}
