% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIPALS_PCAcomponent.R
\name{NIPALS_PCAcomponent}
\alias{NIPALS_PCAcomponent}
\title{Calculating a single PCA component}
\usage{
NIPALS_PCAcomponent(X)
}
\arguments{
\item{X}{Input matrix with rows and columns representing observations and variables.}
}
\value{
Returned is a list with the following entries:
\item{Residual X}{Residual X matrix.}
\item{Scores}{PCA component scores.}
\item{Loadings}{PLS component loadings.}
}
\description{
Calculating a single PCA component
}
\references{
{Geladi, P and Kowalski, B.R. (1986), Partial least squares and regression: a tutorial. Analytica Chimica Acta 185, 1-17.}
}
