% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIPALS_PLS_component.R
\name{NIPALS_PLS_component}
\alias{NIPALS_PLS_component}
\title{Calculating a single PLS component}
\usage{
NIPALS_PLS_component(X, Y)
}
\arguments{
\item{X}{Input matrix with rows and columns representing observations and variables}

\item{Y}{Dependend variable, in form of dummy matrix (multi-levels allowed) or numeric column vector}
}
\value{
Returned is a list with the following entries:
\item{X.res}{Residual X matrix.}
\item{Y.res}{Residual Y matrix.}
\item{scores}{PLS component scores.}
\item{loadings}{PLS component loadings.}
\item{weights}{PLS component weights.}
\item{betas}{PLS X coefficients.}
\item{Q.pc}{PLS Y coefficient.}
}
\description{
Calculating a single PLS component
}
