% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise.est.R
\name{noise.est}
\alias{noise.est}
\title{Estimation of noise level}
\usage{
noise.est(NMR, ppm, where = c(14.6, 14.7))
}
\arguments{
\item{NMR}{Input matrix with rows repesenting sepctra.}

\item{ppm}{ppm vector with its length equals to \code{ncol(X)}.}

\item{where}{Signal free region across all NMR spectra (see Details).}
}
\value{
Returned is a vector of noise levels for each spectrum.
}
\description{
Estimation of noise level
}
\details{
Estiamation of noise level in NMR spectra. This is useful for quality control checks (e.g., before and after spectral normalisation). Noise estimation requires a singal-free ppm region across all spectra, usually this is at the extreme ends or the spectrum. This function requires a minimum number of 50 data points to robustly estimate noise levels.
}
\seealso{
\code{\link{readBruker}}
}
\author{
Torben Kimhofer \email{tkimhofer@gmail.com}
}
