% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca}
\alias{pca}
\title{Perform Principal Component Analysis}
\usage{
pca(X, pc = 2, scale = "UV", center = T)
}
\arguments{
\item{X}{Numeric input matrix with each row representing an observation and each column a metabolic feature.}

\item{pc}{Desired number of principal components.}

\item{scale}{Desired scaling, currently only unit variance scaling (or no scaling \code{scale=FALSE} is implemented.}

\item{center}{Logical indicating if data should be mean centred.}
}
\value{
This function returns a \emph{PCA_MetaboMate} S4 object.
}
\description{
This function is used to perform Principal Component Analysis (PCA).
}
\references{
Geladi, P and Kowalski, B.R. (1986), Partial least squares and regression: a tutorial. \emph{Analytica Chimica Acta}, 185, 1-17.
}
\seealso{
\code{\link{PCA_MetaboMate-class}}
}
\author{
Torben Kimhofer
}
