% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBruker.R
\name{readBruker}
\alias{readBruker}
\title{Reading 1D NMR spectra from Bruker}
\usage{
readBruker(path, filter = T)
}
\arguments{
\item{path}{Path to experiment folder.}

\item{filter}{Logical, if set TRUE then experiments with matching \emph{acqus}, \emph{procs} and \emph{1r} files are read-in. If set FALSE, the algorihm will stop if file systems are corrupt or 2D experiments are presenr. The latter option is for more for running file system tests.}
}
\value{
The following variables are automatically returned as globally defined variables:
\item{X}{Imported NMR matrix where NMR spectra in rows and ppm variables in columns.}
\item{ppm}{Ppm vector.}
\item{meta}{Spectrometer metadata extracted from acqus and procs files (eg time of spectral acquisition).}
}
\description{
Reading 1D NMR spectra from Bruker. This function is recurisvely searching for matching processing, acquisition and spectrum data (r1) files in a given experiment folder.
}
\details{
Paremeters with prefix \code{a} or \code{p} were extracted from the \emph{acqus} or \emph{procs} file, respectively. Paremeters without any prefix were calculated on the fly.
}
\author{
Torben Kimhofer, Imperial College London (2017)
}
