% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.quality.R
\name{spec.quality}
\alias{spec.quality}
\title{Calculating spectral quality indices}
\usage{
spec.quality(X, ppm, ppm.noise = c(9.4, 9.5), plot = T)
}
\arguments{
\item{X}{NMR matrix, rows represent spectra.}

\item{ppm}{Matching ppm vector.}

\item{ppm.noise}{region for noise estimation, must be signal free across all spectra.}

\item{plot}{Logical indicating if a summary plot should be produced.}
}
\value{
Returned is a data frame with TSP line width, residual water signal, normalised baseline estimation (the lower the less baseline fluctuations), estimated signal to noise ratio.
}
\description{
Calculating spectral quality indices
}
\references{
Eilers, P.H.C. (2004), Parametric Time Warping, \emph{Analytical Chemistry}, 76.2, 404–411.

Bloemberg, T.G., \emph{et al.} (2010), Improved parametric time warping for Proteomics, \emph{Chemometrics and Intelligent Laboratory Systems}, 104.1, 65-74.
}
