% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalArea.R
\name{totalArea}
\alias{totalArea}
\title{Total area normalisation}
\usage{
totalArea(X, add.DilF = "ta.dilf")
}
\arguments{
\item{X}{Input matrix where rows represent samples.}

\item{add.DilF}{Character string for new variable for dilution factor (will be exported as global environment variable). Can be left NULL if dilution factor should not be exported.}
}
\description{
Total area normalisation
}
\author{
Torben Kimhofer \email{tkimhofer@gmail.com}
}
