% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totSS.R
\name{totSS}
\alias{totSS}
\title{Calculating Total Sum of Squares (TSS)}
\usage{
totSS(X)
}
\arguments{
\item{X}{Mean centered and unit variance scaled matrix (see Details).}
}
\value{
Total sum of squares.
}
\description{
Calculating Total Sum of Squares (TSS)
}
\details{
Input matrix rows correspond to n observations and columns to m variables. For large matrices (here defined by m>3k), X is subsampled to speed up computation time. Currently, subsampling is achieved by selecting each selecting each third row element.
}
