% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center_scale.R
\name{center_scale}
\alias{center_scale}
\title{Centering and scaleing vectors or matrices}
\usage{
center_scale(X, idc = "all", center = T, scale = c("None", "UV",
  "Pareto"))
}
\arguments{
\item{X}{Data to be scaled, can be a vector or a matrix with observations in rows variables in columns.}

\item{idc}{Indices of samples in X used to derive center/scale parameters (relevant for statistical validation procedures, e.g., scaling X with training set samples). If set to \code{all}, no selection is performed.}

\item{center}{Logical indicating if data should be mean centered.}

\item{scale}{Scaling method that should be applied: none, unit variance (UV) or pareto scaling.}
}
\value{
Centred/scaled data of the same dimensions than input.
}
\description{
Centering and scaleing vectors or matrices
}
