% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca}
\alias{pca}
\title{Perform Principal Component Analysis}
\usage{
pca(X, pc = 2, scale = "UV", center = T, method = "nipals")
}
\arguments{
\item{X}{Numeric input matrix with each row representing an observation and each column a metabolic feature.}

\item{pc}{Desired number of principal components.}

\item{scale}{Desired scaling method: \code{None}, \code{UV} (unit variance) or \code{Pareto} (Pareto scaling).}

\item{center}{Logical indicating if data should be mean centered.}

\item{method}{Algorithm for computing PCA. NIPALS is standard and usually fine. It can handle small amounts of missing/NA values.}
}
\value{
This function returns a \emph{PCA_MetaboMate} S4 object.
}
\description{
This function is used to perform Principal Component Analysis (PCA).
}
\details{
Other methods include: 'svd', 'rnipals', 'bpca', 'ppca', 'svdImpute', 'robustPca', 'nlpca', 'llsImpute', 'llsImputeAll'. If these methods are specified, the \code{pca} function from the \code{pcaMethods} package is used to fit PCA model (see References).
}
\references{
Geladi, P and Kowalski, B.R. (1986), Partial least squares and regression: a tutorial. \emph{Analytica Chimica Acta}, 185, 1-17.
}
\seealso{
\code{\link[=PCA_MetaboMate-class]{PCA_MetaboMate}} \code{\link[pcaMethods]{pca}} \code{\link{plotscores}} \code{\link{plotload}} \code{\link{opls}}
}
\author{
Torben Kimhofer \email{tkimhofer@gmail.com}
}
