% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBruker.R
\name{readBruker}
\alias{readBruker}
\title{Importing 1D NMR spectra from Bruker}
\usage{
readBruker(path, filter = T)
}
\arguments{
\item{path}{A character string, specifying the path of the experiment folder location.}

\item{filter}{Logical, if set to TRUE then experiments with matching \emph{acqus}, \emph{procs} and \emph{1r} files are read-in, ignored are corrupt file systems and 2D experiments. If set to FALSE, the algorihm will be abborted if a file system is corrupt or if 2D experiments are present. Generally the filter=TRUE option should be preferred unless for some testing operations.}
}
\value{
The following variables are automatically returned as globally defined variables:
\item{X}{two-dimensional NMR matrix with spectra in rows and chmemical shift variables in columns.}
\item{ppm}{chemical shift vector (ppm).}
\item{meta}{spectrometer metadata extracted from acqus and procs files (eg time of spectral acquisition).}
}
\description{
Reading-in 1D NMR spectra in Bruker file format. The function recursively descends the directory tree of a given location and is searching for corresponding processing (\emph{procs}), acquisition (\emph{acqus}) and spectrum (\emph{r1}) files.
}
\details{
Columns in meta represent spectrometer meta-vairables. Prefix \code{a} and \code{p} indicated if a parameter was extracted from the \emph{acqus} or \emph{procs} file, respectively. Paremeters without any prefix were calculated on the fly (eg run-order).
}
\author{
Torben Kimhofer \email{tkimhofer@gmail.com}
}
